Move setup command to their own file
